+++ inject/inject-app/src/test/scala/com/twitter/inject/app/StartupTimeoutException.scala
@@ -1,3 +1,8 @@
 package com.twitter.inject.app
 
-class StartupTimeoutException(message: String) extends Exception(message)
+class StartupTimeoutException(message: String, cause: Throwable) extends Exception(message, cause) {
+
+  def this(message: String) {
+    this(message, null)
+  }
+}